Move key related logic to token module
